Fix medium code-quality issues: ViewModel, R8, NCS, dead code
- Delete JamarrPlaybackContext.kt (dead code, only written to, never read)
- Extract JamarrViewModel from JamarrApp.kt (622 → 391 lines)
- Enable R8 minification/shrinking in release builds
- Add Network Security Config (cleartext blocked in release)
- Replace hardcoded server URL with BuildConfig field
- Delete PlaceholderScreen.kt (dead code, never referenced)
- Extract shared formatMs utility, dedupe in NowPlayingScreen/MiniPlayer
