--- android/app/src/main/java/com/jamarr/android/playback/JamarrPlaybackContext.kt
@@ -1,11 +1,0 @@
-package com.jamarr.android.playback
-
-import com.jamarr.android.data.JamarrApiClient
-
-object JamarrPlaybackContext {
-    @Volatile
-    var serverUrl: String = ""
-
-    @Volatile
-    var apiClient: JamarrApiClient? = null
-}
