+++ android/app/src/main/java/com/jamarr/android/util/TimeUtils.kt
@@ -1,0 +1,8 @@
+package com.jamarr.android.util
+
+fun formatMs(ms: Long): String {
+    val totalSeconds = (ms / 1000).coerceAtLeast(0)
+    val minutes = totalSeconds / 60
+    val seconds = totalSeconds % 60
+    return "%d:%02d".format(minutes, seconds)
+}
